Update patch set 1
Patch Set 1:

Adding it as fields to the GlobalConfig makes sense to me.

It would be "project: chromeos-status" at console.cloud.google.com. I don't have permission to add you to the project, unfortunately. You need to find the owner and get them to add yourself and chrome-troopers@ as Owners or Editors.

You don't need to do anything to the DB other than adding the fields to the db.Model class.

Patch-set: 1
Label: Code-Review=0