[som] Fix some frontend tests and code broken by the reasons refactoring.
In the future we need to be more careful about making sure that changes
to the JSON format are tested on the client and don't break things there.
The example alerts in test/alerts-data.js should be updated whenever we
make changes to alerts-dispatcher that alter the structure of the json it
produces.

BUG=640879

Change-Id: Ie515ad81bcfe7fdc3ab1a48d515fff176be0c0a6