Update patch set 1
Patch Set 1:

We accidentally deleted the bug queue label data from staging at some point (probably when adding the display names). 

This was because the trees are all deleted when the tree field is submitted and the bug queue labels field defaulted to blank. 

(I imagine this issue would also happen for alertStreams, but it is a little less straightforward to resolve there because aleretStreams' input format appears to only take in one tree at a time)

Patch-set: 1
Label: Code-Review=0