Update patch set 1
Patch Set 1:

(10 comments)

Thanks a lot for digging all this info and writing it down. I am sure it will come useful in the future.

Overall, this looks more like a story than documentation. I am fine with this, but this is different from docs everywhere else. :-)

Also please try to be consistent using `x` for specifying inline snippets of configs or code and for specifying well-known file names. I prefer it to the use of "x" and especially to `"x"`, but IMHO this is up to your personal style.

Patch-set: 1