swarmbucket: make canarying non-deterministic
currently swarmbucket deterministically decides whether the canary task
template must be used for a build. Although elegant in theory, in
practice it meant that if canary is broken and a CL was unfortunate
enough to use the canary, all CQ's retries on the build will fail. A
human will have to submit to CQ again and that's bad.

Make canarying non-deterministic. Now if a canary is broken and was used
for first build attempt for a CL, a non-canary template may be used for
the retry, so users won't be annoyed that infrastructure does not let
their CLs through. Luci team will still see discrepancy between canary
and non-canary templates.

R=vadimsh@chromium.org, estaab@chromium.org
BUG=

Change-Id: I3c22439e83e5aca84f3e6a6fc19b6b62cfa1e04c