kitchen: pass cache and temp dir
Add a flag for temp dir. If a flag for cache/temp directory is passed,
propagate it to the recipe_engine/path module via a property..

We will pass the flags in swarming task template. The result is that we
control paths the recipes will use. In particular, we will map "git"
named cache to the [CACHE]/git recipe path and bot_update will use it.
Also it makes sure that recipe engine uses <rundir>/tmp directory and
not system temporary dir, because system temp dir may be on a smaller
drive and because the rundir is temporary by definition (Swarming
deletes it after task completion)

R=estaab@chromium.org
BUG=662586

Change-Id: I7fa0ea94c9813a5ce7b2f87f14311237cb8f4afd