Reduce number of paths added to sys.path
This will reduce number of conflicts created by all the added paths. In
particular it allows to add test folder into the root of the chromium_try_flakes
app, which will contain tests for main module. Previously it was conflicting
with dateutil/test.py module, but now we only symlink the dateutil/dateutil
folder.

This is also helpful in preparation before moving all sys.path modifications to
files that need them, which will allow to run chromium_try_flakes tests in
parallel despite the multiprocessing import issues in expect_tests.

BUG=667586

Change-Id: I7536b846c8b68aa68dd7352964a40fbd76eabc09