Update patch set 1
Patch Set 1:

martiniss@ mentioned in our OKR meeting that it was difficult to tell when there was a comment to look at, so I decided to test changing the UI to change the color of the comments link when there was a comment. 

It was a bit difficult to find a good color for this. I ended up just going with the FindIt color since I don't want to add too many random colors to the UI. 

Demo: https://7477-10eec3a-tainted-zhangtiff-dot-sheriff-o-matic-staging.appspot.com/chromium

Small thing. Eventually, I think I might want to move where comments are located in the UI, but I'm not sure what a better place to be would be.

Patch-set: 1