Update patch set 2
Patch Set 2:

My theory right now is that these alerts are usually not fired for overprovisioned reasons, but usually for when cache in /b/folder_a gets moved to /b/folder_b and it didn't get cleaned up.

But yeah it is important not to lump this in with overprovisioned events... I'm not sure how to differentiate it.

Patch-set: 2