Update patch set 2
Patch Set 2:

> Patch Set 2:
> 
> > Patch Set 2:
> > 
> > My theory right now is that these alerts are usually not fired for overprovisioned reasons, but usually for when cache in /b/folder_a gets moved to /b/folder_b and it didn't get cleaned up.
> > 
> > But yeah it is important not to lump this in with overprovisioned events... I'm not sure how to differentiate it.
> 
> A better effort that will handle this was started here: https://chrome-internal-review.googlesource.com/c/321168/ and https://chrome-internal-review.googlesource.com/c/321207/
> 
> To finish it, we need to start putting 'slavebuilddir' (or whatever it is called) into botmap and make sure it propagates all the way down to cleanup_disk.

Or, alternatively, we can teach buildbot to update some timestamp-like file when it touches a cache dir, and then cleanup_disk can use it to decides when to nuke the cache dir completely.

In any case, I believe we should kill only stale caches (and report this as a metric!), not all of them at once.

But whatever.

Patch-set: 2