Update patch set 1
Patch Set 1: Code-Review+1

(2 comments)

Failing the build doesn't always mean that someone will look at it. gatekeeper on alerts on a change of status; it will email you when the build starts to fail, but it will not continue to email you when the build has been failing for a while. This is a good step, but I think we'll want more eventually.

Patch-set: 1
Reviewer: Gerrit User 1148599 <1148599@3ce6091f-6c88-37e8-8c75-72f92ae8dfba>
Label: Code-Review=+1