[Findit] Flake Analyzer - Fix infinitely running tryjobs
The root cause is the case where a flaky test is newly added is not accounted for correctly. The lookback algorithm should check for such a case and return the revision that adds the flaky test as the culprit, rather than rely on the calling code to check for next_run_number and realize the run number has already been analyzed and complete accordingly.

BUG=694748

Change-Id: I2401035cfdc7bd2e3cf12d03a8769ccf6ed55c2f