[Predator] Use linear feature value instead of log linear value.
Previous, a feature value is in (log(0), log(1)], for example, feature like ``MinDistanceFeature``, if distance between changed lines and crashed lines are more than 50, we score it log(0), in this case, no matter what value it gets for other features, we won't consider this suspect as culprit. 

However, this way of evaluating features will make the log linear model hard to scale. For example, even thought a suspect didn't touch crashed files in a stacktrace, we still want to checkout whether it touched files under the same directory of crashed files. So we need to switch the value range from (log(0), log(1)] to [0, 1].
 

BUG=695619

Change-Id: I3c4f0400b08867540d3d96098cd458435aa2df45