[Findit] Properly serialize datetime callback parameter
The parameter start_time when defined is a datetime object which cannot
be serialized with json.dumps. This change converts it to a string using
the functions exposed by time_util before serializing it and back to
datetime when deserializing.

Also, adding data to the existing test to force this behavior.

R=chanli,stgao
BUG=703761

Change-Id: I40bc1e744806f2c4aadf0ce6609aaa61b4019fa7