Update patch set 7
Patch Set 7:

The code style lgtm. I'm not sure what your future plans for this app are, but I do want to make an overall comment that it might make sense to structure this as JavaScript which requests some sort of Go API (probably an endpoint that returns just the JSON you want) and then sets the view of the page according to the data. 

This would mainly be for if you eventually wanted to have the JavaScript autorefresh the content on the page every now and then.

Patch-set: 7
Reviewer: Gerrit User 1159599 <1159599@3ce6091f-6c88-37e8-8c75-72f92ae8dfba>