Remove /data/ from revision range URLs on Flakiness Dashboard
As part of moving JSON endpoints on Test Results Server into /data/ subpath,
I've also moved /revision_range to /data/revision_range. Later I have realized
that this is not a JSON endpoint and therefore moved it back to /revision_range.
However, I forgot to also revert changes to the URLs in the Flakiness Dashboard
UI, which is what this CL fixes.

Bug: 712805
Change-Id: Id3f02b1b57980b83ca70c246b18d7c75c296cb07