git: Add carat escaping to Windows batch.
If the Git wrapper invokes Git using a batch script on Windows, any
carats in the command line will be regarded by the batch script as
escape characters, and need to be escaped. Have the Git wrapper
automatically do this when (a) running on Windows, and (b) running Git
through a ".bat" file.

BUG=None
TEST=local
  - Before patch: git rev-parse HEAD^ (shows HEAD)
  - After patch: git rev-parse HEAD^ (shows HEAD^)

Change-Id: I62da677c56fe2a5e304ecf5f0e0ec8cfad36ac13