[bot_setup] Remove files as root when possible.
"bot_setup" operates with the assumption that it, as a non-root user,
can recurisvely delete certain files. Unfortunately, this no
longer holds on ChromeOS bots, which install "chroot". Update the script
to try and delete files as root, falling back to non-root when it
cannot.

To facilitate this, introduce "infra.tools.rmtree", a script which
recursively deletes files and knows how to escalate permissions.

BUG=chromium:734674
TEST=local
  - Tested against root-owned files.
  - Tested against non-root-owned files.
  - Tested after `sudo -k` to clear the sudo timestamp, falls back to
    non-root deletion.

R=hinoka@chromium.org

Change-Id: I36745cb8107e3a8dc148b5b58f7737cc0027330e