[cr-audit-commits] Defer unset clients.
Without this change, the cron jobs are running into "invalid security
ticket" errors. It seems that the RepoConfig is persisting between cron
job runs, perhaps because it's contained by the RulesMap global.

This change unsets the clients when the handler completes its run so
that in case the global is reused, the new request makes new clients.

R=vadimsh,iannucci,stgao
BUG=764445

Change-Id: I097dc342abd01b39e2bb3a5090b6f78e29b01ad6