Update patch set 5
Patch Set 5:

> Patch Set 5:
> 
> Can you de-dupe the multiple selection users? https://screenshot.googleplex.com/Jbvx3aYPPDO

Context: The plan is to have the parent element update 'suggestions' whenever the input and selected users change.
So, when a user is selected, 'suggestions' should be updated to not include the user that was just added and it should be impossible to select the same user twice.

I did this to avoid calculating the 'suggestions' and leave it up to whomever uses this element to do it. Also, I thought it might be really slow to have user-input take ALL possible users.

What do you think of strategy in general and in the context of your de-duping comment?

Patch-set: 5
Reviewer: Gerrit User 1167775 <1167775@3ce6091f-6c88-37e8-8c75-72f92ae8dfba>