[Findit] Taskqueue - In memory representation of the taskqueue.
This is the in-memory representation of the taskqueue which polls
existing swarming_task_queue_requests that exist in the datastore
and arranges then in separate queues.

Dimensions mutliqueue - Tracks all requests that use a certain
configuration.
Pending queue - Tracks all requests that have been sent to swarming.
Completed queue - Tracks all requests that have been completed, and 
are waiting for results to be queried and returned to the calling 
pipeline.

Full design here: go/findit-scale-up-flake-analyzer

Bug:775186
Change-Id: Ic915258c46972a8d09a1df1b813ef9cab6f1679c