[buildbucket] add build.experimental attribute
We need to schedule experimental builds that do not make side effects and that
are by default invisible, e.g. not surfaced in dashboards, monitoring, search
results.

Add experimental bit to build, accept it on build creation, or at builder level.
Also integrate it with luci-migration and derive
'$recipe_engine/runtime'.is_experimental property from this attribute.

This slightly changes behavior of the integration with migraiton app.
Was: if migration host is specified, by default a build is
experimental. Now: default depends on the builder.experimental config.

This CL does not update search or monitoring. It will be done in next CLs.

Bug: 787089
Change-Id: I9aa694ad0fbdae8a5729fbc00b379a46bfce407f