Update patch set 1
Patch Set 1:

Though I just realized that if we want to keep chops-ui, then we can't rename the chopsui repo because the Bower registration is hardcoded to the repo URL and unchangeable. So if we rename everything it will probably be to chopsui.

Patch-set: 1
Reviewer: Gerrit User 1159599 <1159599@3ce6091f-6c88-37e8-8c75-72f92ae8dfba>