Update patch set 2
Patch Set 2:

Here's the schema so far for charting. This works for my prototyping but I'm not 100% confident it won't need to be changed in some way in the future.

For this CL - I'd imagine we'd only land it once it's 100% ready to go into production. With that in mind, would the right next step be to test this CL on the staging database?

Patch-set: 2
Reviewer: Gerrit User 1163684 <1163684@3ce6091f-6c88-37e8-8c75-72f92ae8dfba>
Reviewer: Gerrit User 1168241 <1168241@3ce6091f-6c88-37e8-8c75-72f92ae8dfba>
Work-in-progress: false