[Monorail] Make notify tasks use comment_id.  CL 1 of 3.
In this CL, I am changing all the easy-to-change callers of
PrepareAndSendIssueChangeNotification() and making
NotifyIssueChangeTask() handle both the old and new set of parameters.

This part alone should fix the problem so that users have no chance
of being notified of the wrong comment, even if there is replication lag.

The other two parts will make our code cleaner and more consistent.  In
part 2, I will update the other callers, which will require a bit more
refacotring.  And, in part 3 in a later deployed version, I will make
comment_id required and drop the seq_num approach.

R=jojwang@chromium.org

Bug: monorail:3550
Change-Id: I3a266a85e1f3201e2bf72023e70678a186df320f