Update patch set 2
Patch Set 2:

(1 comment)

Wanted to stop here and get your feedback on this approach before going any further.

At a high level my idea was to add a "snapshot mode" to ast2select.py that switches it from using the Issue table to using the IssueSnapshot table. For the conditions we can't support with snapshots, it returns that condition as unsupported. The list of unsupported conditions is returned in the JSON API response to be displayed to the user.

This currently works for basic queries. Lmk if there's a better way.

Patch-set: 2
Reviewer: Gerrit User 1163684 <1163684@3ce6091f-6c88-37e8-8c75-72f92ae8dfba>
Reviewer: Gerrit User 1168241 <1168241@3ce6091f-6c88-37e8-8c75-72f92ae8dfba>
Work-in-progress: false