Update patch set 2
Patch Set 2:

> Patch Set 1:
> 
> I think ignoring types and converting everything to string will cause some problems down the line. Do you know for a fact that in requests send to swarming this is given as a string?

It appears we can send either a string or an int to Swarming, but when specifically getting json responses from Swarming when getting a task request, it specifically returns each int field as a string. Honestly I'm not a fan either but it's better we keep the layer that talks to swarming as consistent as possible.

Patch-set: 2