Update patch set 2
Patch Set 2:

(3 comments)

Yeah, I think this is the approach that I had in mind.  It is adding some complexity to a part of the code that is already pretty tricky, but I think that it will definitely work.

Patch-set: 2
Reviewer: Gerrit User 1163684 <1163684@3ce6091f-6c88-37e8-8c75-72f92ae8dfba>