[monorail] Add some monitoring for garbage collection
This measures the counts of objects in each of the GC's generations.

It *attempts* to count reuquests where at least one GC event occurred.

The latter is probably an undercount but there's no obvious way to
determine if a GC event occurred over some period of time other than
comparing the generation sizes before and after.

Change-Id: I562ec0f4c81f80aee5e62ac646f8e2803bcda4b6