[Findit] Flake Analyzer - Fixing running existing data points
The root cause is a corner case when exponential search lands exactly
on a data point that has already been ran, when the distance is exactly
a square number. There is an off-by-1 error checking for when to restart
exponential search from 1, causing the existing data point to be suggested
to run again, which causes problems later on in the analysis.

1. Fix the off-by-1 error
2. Add assertion checks to ensure existing data points are never suggested
   again
3. Add assertion checks for when a suspected build is identified to ensure
   it is indeed a suspect (stable -> flaky).

Bug: 836976
Change-Id: I68e0f8014a9efe8e01009bd380a9d255a9806941