[Findit] Flake Analyzer - Round off commit position to nearest builds
Previously, requested commit positions were mapped to the nearest valid
builds on either side. next_commit_position_utils would then try to
select the later of the two builds first in most cases, falling back to
the earlier build if the later of the two was already analyzed. This leads
to cases where if a requested commit position has a large step size from
the previous, but the bounding builds range is large, the net step size
ends up being small. The net effect is exponential search might degenerate
into almost linear search.

The new approach is to round off, and use the commit position of the closer
build to the requested commit position, not just always using the more recent
build all the time.

Bug: 846402
Change-Id: Ib93d23b346247ac26c262ecaf5bc3ba357db24ad