Update patch set 1
Patch Set 1:

Do you still want the code to have a limit?
Or, is it fine to be unlimited?

If you prefer to limit the size, is 12 acceptable?
If yes, I am ok with that for now, and I will update this cl.

Patch-set: 1
Reviewer: Gerrit User 1107806 <1107806@3ce6091f-6c88-37e8-8c75-72f92ae8dfba>