metrics: increase limit from 7 to 12.
There seems not be limit for the number of metric fields according to
the manual, and limiting the number of metric fields make me write
another data source to represents the same data with buildbot, which
does not make any sense.  Thus, my original goal was removing the limit.

At the same time, the author of the library has a model that people put
two possible values to each fields.  To avoid hitting limit of max cardinality
with the scenario, 12 fields is the max. (2^12 = 4096)
Note maximum cardinality is 5000 (see http://shortn/_WBupjZf2of).
In previous model, 5 built-in fields are included in the calcuration,
and 12 - 5 = 7 was the limit.  However, the build-in fields have only
one combination of values, and it is irrelvant to include them to calculate
cardinality limit.

Since my goal is adding BuilderId (project, bucket, name), having 3 more
metrics is enough.  Increasing limit to 12 instead of removing the limit
should be reasonable point of compromise.

Please see also discussion in:
https://chromium-review.googlesource.com/c/infra/infra/packages/infra_libs/+/1068578

Bug: b/78316701
Change-Id: I63132aee294cee0ea56f23d0e1850ea36548c39f