publish_tarball: rebuilt lite, nacl, and testdata tarballs if necessary
publsh_tarball used to only check if the full tarball was published, and if it
was, it would assume the lite, nacl, and testdata ones were published too.  This
assumption was recently violated since the nacl tarballs for specific versions
weren't getting published (since the bot was reaching its timeout and the nacl
tarballs get pushlished last).  The new behavior in this CL is to check each
tarball individually.

BUG=847213
R=mmoss

Change-Id: I476805da5a44c216e33d2cc953c1e189776645f3