[Findit] Flake Analyzer - Set analysis completion status at correct time
Analysis status is being set even when there are more subpipelines. When
one of those subpipelines fails, the logic to check for updating the
status will bail out if the status is set to COMPLETED thus the errors
are not obvious.

This change is to set the analysis status only after all subpipelines
have completed.

Bug: 847641
Change-Id: I5900d1850bf1618283ca01d2a449e4e4cd1d0b84