[buildbucket] Remove tag index order requirement
TagIndex entires are currently guaranteed to be sorted.

One TagIndex entity does not scale to the QPS we have today for a same
buildset. It needs to be sharded. To continue benefiting from tag index entry
ordering, we'd need to implement merging of sorted arrays.
It is not worth it. We have up to 1000 entries. It takes <1ms to sort them
on macbookpro

R=vadimsh@chromium.org

Bug: 838318
Change-Id: I4362a420b730d1e8d12252677f5fb3b81554c3cd