Update patch set 8
Patch Set 8:

(4 comments)

> Patch Set 7:
> 
> > Patch Set 7:
> > 
> > Sean, what do you think about using Polymer 3 from a high level?
> > 
> > In theory, I think this makes the most sense if there's some chance that the Chops frontends will migrate to Polymer 3 later. But, even if not though, it seems OK for this case.
> > 
> > Sean, Marc-Antoine, Tiffany, Kevin, free to also take a look over to point out any potential concerns or improvements here.
> 
> FYI regarding Polymer 3: At some point, we need to decide whether we want to use Polymer 3 or skip directly to web components using lit-element (ie: https://github.com/Polymer/pwa-starter-kit/ )
> 
> I've only briefly looked at the options, but I'm leaning towards using lit-element, especially since Chops users tend to have modern browsers. Though with lit-element, I think ChopsUI's Polymer 2.0 components wouldn't work unless you added Polymer as well. I do want to rewrite the more used ChopsUI elements as web components soon, but I can't promise a solid timeline on that. If that's liveable, I think we should lean lit-element if we can make it work nicely.

Makes sense. Looking at the pwa-starter-kit project, the development tools (polymer-cli etc) and package management is similar, so for a simple app like this one, the difference between using Polymer 3 and lit-element should be pretty small, and it should be easy to change from one to the other. (And conveniently for future migration, this could be done one element at a time.)

I think I could upload a version of this CL with lit-element by EOD so that we could compare, if that would be useful.

Patch-set: 8
Reviewer: Gerrit User 1003190 <1003190@3ce6091f-6c88-37e8-8c75-72f92ae8dfba>