[cipd] Disable ndb memcache to allow Go backend to run side-by-side.
Python's ndb and Go's dscache are not compatible with each other: an entity
written by Python side may be read as stale from Go and vice versa.

Disable the caching completely for the duration of v1 => v2 transition, when
both Python and Go backends run concurrently on top of same data.

R=nodir@chromium.org, tandrii@chromium.org
BUG=801423

Change-Id: If65321215186146b60eb4dcedec8bba0f888f74e