[tricium] Re-enqueue the collect request when the task is incomplete.
Related to https://crrev.com/c/1123062.

Previously when a lack of isolated output meant the task
was considered failed, it was retried until it was successful.

It appeared that there were two pubsub notify requests so
ignoring the first would be OK, but after deploying to
tricium-dev I saw that actually it seems to result in tasks
that never get their output collected and never get marked
complete.

This CL would make it so that when a collect request is
made but finds no isolated output, it retries repeatedly
until success.

Bug: 1123062
Change-Id: I4dd2525a0ddbae4ab71954bb35b42989fea5c1df