[Findit] Flake Analyzer - GetIsolateShaPipeline to use new build indexing
One of several changes needed to phase out step_util.GetValidBoundingBuildsForStep,
which is the weakest point in Flake Analyzer. This change is to switch to using
IsolatedTarget's methods to relevant builds by commit position, which should be
significantly more robust. Builds not indexed are assumed not supported, as they
should all be indexed in recent history.

Bug: 850224
Change-Id: I5bedd3f22137ce983bd025d26a752de1edab39b5