Update patch set 5
Patch Set 5:

Requires https://chromium-review.googlesource.com/1158605 to be rolled in first.

I've made is somewhat limited, to the sense that only one cache entry can be marked as a fallback for the moment. I did so for logic simplicity and reduce the chance of surprising behavior, but I'm fine to adjust to whatever you think is better.

Patch-set: 5
Reviewer: Gerrit User 1000487 <1000487@3ce6091f-6c88-37e8-8c75-72f92ae8dfba>
Work-in-progress: false