Update patch set 5
Patch Set 5:

I know you're working on some adjustments to 3pp, but I'd love to be unblocked on moving forward with this. This now properly packages the node_modules subdir under nodejs with firebase-tools in it -- I think that should be sufficient for me to use it?

Patch-set: 5
Reviewer: Gerrit User 1271098 <1271098@3ce6091f-6c88-37e8-8c75-72f92ae8dfba>