Update patch set 2
Patch Set 2:

(1 comment)

Does this change (and the previous code) mean that if webkit_layout_tests failed, we wouldn't see that on SoM, but if a different test failed (say, base_unittests) we should've?

I'm not positive (I wish SoM had history), but I feel like I did see the webkit_layout_tests results once, just not every time. So I'm not sure if I think this really solves the problem.

But, we can try it and see; it seems like it should at least be an improvement.

Patch-set: 2