[tricium spellchecker] Case-insensitive flagging and suggestions
This change would allow flagging misspellings regardless of case.
This is done by setting all dictionary keys to lower-case, and
lower-casing words when looking up in the dictinoary.

In order to give better suggestions, however, the case of the
proposed fix should ideally match the mispelling; for example,
if the misspelling was "Occurences", then the proposed fix
should be "Occurrences", not "occurrences" -- so, this CL also
converts the list of fixes to match the misspelling.

Bug: 873311
Change-Id: Ie00117ca38cd53cf7331a25a367a33931c394a10