[tricium spellchecker] Case-insensitive flagging and suggestions
This change would allow flagging misspellings regardless of case.
This is done by ensuring all dictionary keys are lower-case, and
words to check are lower-cased when looking up in the dictionary.

In order to give better suggestions, however, the case of the
proposed fix should ideally match the mispelling; for example,
if the misspelling was "Occurences", then the proposed fix
should be "Occurrences", not "occurrences".

Bug: 873311
Change-Id: Ie00117ca38cd53cf7331a25a367a33931c394a10