[cr-buildbucket] fix typo in named cache fallback dimension name
The bots are exposing the named caches as dimension 'caches', not
'cache'.

It didn't break anything, as the Swarming server used immediate fallback
(aka NO_RESOURCE) and fell to the cold cache. This change fixes this and
will make the warm cache work.

It's a fairly safe change, since it's changing the injected dimension
key only for the hot task slice.

R=nodir@chromium.org

Bug: 847602
Change-Id: Iac7a2914379d31cabd7a12b7d4841188f633c5e9