qscheduler: unify mutaters into a single proto
The Mutater interface is unnecessary; all that callers need is a type
that describes tasks being assigned to workers. Create a new proto for
this (Assignment) and drop unnecessary Mutator abstraction.

Internally, we still define an apply method for Assignment, but this is
mainly for convenience, and we do not export it.

BUG=chromium:870537
TEST=unit tests

Change-Id: I76025d1057c5be339a5584801d2ffa335ff8b36a