Update patch set 2
Patch Set 2:

(3 comments)

One high level question. It looks like the backend is not implemented yet, so maybe it's better to hide the UI by default for now? I think it will be frustrating to users if they see this UI and try to use it, but find out it doesn't work at all.

Patch-set: 2